requirements.txt
discord.py
gtts
jishaku